EarthData

A Julia interface to https://search.earthdata.nasa.gov

Installation

] add EarthData

Usage

julia> using EarthData

# Search for all GEDI L2A granules
julia> gg = granules(short_name="GEDI02_A")
1000-element Vector{EarthData.UMM_G}:
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G

# A single granule, with the non-empty fields shown by default
julia> g[1]
EarthData.UMM_G
        TemporalExtent
        CollectionReference
        RelatedUrls
        GranuleUR
        Platforms
        DataGranule
        MetadataSpecification
        PGEVersionClass
        Projects
        AdditionalAttributes
        ProviderDates
        SpatialExtent
        OrbitCalculatedSpatialDomains
        MeasuredParameters

julia> gg[1].RelatedUrls[1].URL
"https://e4ftl01.cr.usgs.gov//GEDI_L1_L2/GEDI/GEDI02_A.002/2019.04.18/GEDI02_A_2019108002012_O01959_01_T03909_02_003_01_V002.h5"

# Search for s3 urls

# Download a file from S3

Next

Implementing collections will be next on the list to implement.

EarthData.AccessConstraintsType

EarthData.AdditionalAttributeType

EarthData.BoundaryType

EarthData.BoundingRectangleType

EarthData.CharacteristicType

EarthData.ChecksumType

EarthData.CollectionReferenceType

EarthData.CollectionRequest

EarthData.DataGranuleType

EarthData.ExclusiveZoneType

EarthData.FilePackageType

EarthData.FileType

EarthData.GPolygonType

EarthData.GeometryType

EarthData.GranuleRequest

EarthData.HorizontalSpatialDomainType

EarthData.IdentifierType

EarthData.InstrumentType

EarthData.LineType

EarthData.MeasuredParameterType

EarthData.MetadataSpecificationType

EarthData.OrbitCalculatedSpatialDomainType

EarthData.OrbitType

EarthData.PGEVersionClassType

EarthData.PlatformType

EarthData.PointType

EarthData.ProjectType

EarthData.ProviderDateType

EarthData.QAFlagsType

EarthData.QAStatsType

EarthData.RangeDateTimeType

EarthData.RelatedUrlType

EarthData.SpatialExtentType

EarthData.TemporalExtentType

EarthData.TilingCoordinateType

EarthData.TilingIdentificationSystemType

EarthData.TrackPassTileType

EarthData.TrackType

EarthData.VerticalSpatialDomainType

EarthData.granules

EarthData.netrc!

EarthData.AccessConstraintsType — Type

Information about any physical constraints for accessing the data set.

EarthData.AdditionalAttributeType — Type

A reference to an additional attribute in the parent collection. The attribute reference may contain a granule specific value that will override the value in the parent collection for this granule. An attribute with the same name must exist in the parent collection.

EarthData.BoundaryType — Type

A boundary is set of points connected by straight lines representing a polygon on the earth. It takes a minimum of three points to make a boundary. Points must be specified in counter-clockwise order and closed (the first and last vertices are the same).

EarthData.BoundingRectangleType — Type

This entity holds the horizontal spatial coverage of a bounding box.

EarthData.CharacteristicType — Type

This entity is used to reference characteristics defined in the parent collection.

EarthData.ChecksumType — Type

Allows the provider to provide a checksum value and checksum algorithm name to allow the user to calculate the checksum.

EarthData.CollectionReferenceType — Type

A reference to a collection metadata record's short name and version, or entry title to which this granule metadata record belongs.

EarthData.CollectionRequest — Type

CollectionRequest(; keyword=value, ...)

As documented by https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html#collection-search-by-parameters.

EarthData.DataGranuleType — Type

This entity stores the basic descriptive characteristics associated with a granule.

EarthData.ExclusiveZoneType — Type

Contains the excluded boundaries from the GPolygon.

EarthData.FilePackageType — Type

This set of elements describes a file package or a file that contains other files. Normally this is either a tar or a zip file.

EarthData.FileType — Type

This set of elements describes a file. The file can be a part of the entire granule or is the granule.

EarthData.GPolygonType — Type

A GPolygon specifies an area on the earth represented by a main boundary with optional boundaries for regions excluded from the main boundary.

EarthData.GeometryType — Type

This entity holds the geometry representing the spatial coverage information of a granule.

EarthData.GranuleRequest — Type

GranuleRequest(; keyword=value, ...)

As documented by https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html#granule-search-by-parameters. See fieldnames(GranuleRequest) for a list of all possible keywords.

EarthData.HorizontalSpatialDomainType — Type

Information about a granule with horizontal spatial coverage.

EarthData.IdentifierType — Type

This entity stores an identifier. If the identifier is part of the enumeration then use it. If the enumeration is 'Other', the provider must specify the identifier's name.

EarthData.InstrumentType — Type

A reference to the device in the parent collection that was used to measure or record data, including direct human observation. In cases where instruments have a single composed of child instrument (sensor) or the instrument and composed of child instrument (sensor) are used synonymously (e.g. AVHRR) the both Instrument and composed of child instrument should be recorded. The child instrument information is represented by child entities. The instrument reference may contain granule specific characteristics and operation modes. These characteristics and modes are not checked against the referenced instrument.

EarthData.LineType — Type

This entity holds the horizontal spatial coverage of a line. A line area contains at lease two points.

EarthData.MeasuredParameterType — Type

This entity contains the name of the geophysical parameter expressed in the data as well as associated quality flags and quality statistics. The quality statistics element contains measures of quality for the granule. The parameters used to set these measures are not preset and will be determined by the data producer. Each set of measures can occur many times either for the granule as a whole or for individual parameters. The quality flags contain the science, operational and automatic quality flags which indicate the overall quality assurance levels of specific parameter values within a granule.

EarthData.MetadataSpecificationType — Type

This object requires any metadata record that is validated by this schema to provide information about the schema.

EarthData.OrbitCalculatedSpatialDomainType — Type

This entity is used to store the characteristics of the orbit calculated spatial domain to include the model name, orbit number, start and stop orbit number, equator crossing date and time, and equator crossing longitude.

EarthData.OrbitType — Type

This entity stores orbital coverage information of the granule. This coverage is an alternative way of expressing granule spatial coverage. This information supports orbital backtrack searching on a granule.

EarthData.PGEVersionClassType — Type

This entity stores basic descriptive characteristics related to the Product Generation Executable associated with a granule.

EarthData.PlatformType — Type

A reference to a platform in the parent collection that is associated with the acquisition of the granule. The platform must exist in the parent collection. For example, Platform types may include (but are not limited to): ADEOS-II, AEM-2, Terra, Aqua, Aura, BALLOONS, BUOYS, C-130, DEM, DMSP-F1,etc.

EarthData.PointType — Type

The longitude and latitude values of a spatially referenced point in degrees.

EarthData.ProjectType — Type

Information describing the scientific endeavor with which the granule is associated.

EarthData.ProviderDateType — Type

Specifies the date and its type that the provider uses for the granule. For Create, Update, and Insert the date is the date that the granule file is created, updated, or inserted into the provider database by the provider. Delete is the date that the CMR should delete the granule metadata record from its repository.

EarthData.QAFlagsType — Type

The quality flags contain the science, operational and automatic quality flags which indicate the overall quality assurance levels of specific parameter values within a granule.

EarthData.QAStatsType — Type

The quality statistics element contains measures of quality for the granule. The parameters used to set these measures are not preset and will be determined by the data producer. Each set of measures can occur many times either for the granule as a whole or for individual parameters.

EarthData.RangeDateTimeType — Type

Stores the data acquisition start and end date/time for a granule.

EarthData.RelatedUrlType — Type

This entity holds all types of online URL associated with the granule such as guide document or ordering site etc.

EarthData.SpatialExtentType — Type

This class contains attributes which describe the spatial extent of a granule. Spatial Extent includes any or all of Granule Localities, Horizontal Spatial Domain, and Vertical Spatial Domain.

EarthData.TemporalExtentType — Type

Information which describes the temporal extent of a specific granule.

EarthData.TilingCoordinateType — Type

Defines the minimum and maximum value for one dimension of a two dimensional coordinate system.

EarthData.TilingIdentificationSystemType — Type

This entity stores the tiling identification system for the granule. The tiling identification system information is an alternative way to express granule's spatial coverage based on a certain two dimensional coordinate system defined by the providers. The name must match the name in the parent collection.

EarthData.TrackPassTileType — Type

This element stores a track pass and its tile information. It will allow a user to search by pass number and their tiles that are contained with in a cycle number. While trying to keep this generic for all to use, this comes from a SWOT requirement where a pass represents a 1/2 orbit. This element will then hold a list of 1/2 orbits and their tiles that together represent the granules spatial extent.

EarthData.TrackType — Type

This element stores track information of the granule. Track information is used to allow a user to search for granules whose spatial extent is based on an orbital cycle, pass, and tile mapping. Though it is derived from the SWOT mission requirements, it is intended that this element type be generic enough so that other missions can make use of it. While track information is a type of spatial domain, it is expected that the metadata provider will provide geometry information that matches the spatial extent of the track information.

EarthData.VerticalSpatialDomainType — Type

This entity contains the type and value for the granule's vertical spatial domain.

EarthData.granules — Method

granules(; keyword=value, ...) -> Vector{Granule}

Search for granules using NASA EarthDataSearch. Possible keywords are (:collection_concept_id, :granule_ur, :readable_granule_name, :online_only, :downloadable, :browsable, :attribute, :polygon, :equator_crossing_longitude, :equator_crossing_date, :updated_since, :revision_date, :created_at, :production_date, :cloud_cover, :platform, :instrument, :sensor, :project, :concept_id, :echo_granule_id, :echo_collection_id, :day_night_flag, :two_d_coordinate_system, :provider, :native_id, :short_name, :version, :entry_title, :entry_id, :temporal, :cycle, :passes, :sort_key) or (:page_size, :page_num, :offset, :scroll, :sort_key, :pretty, :token, :echo_compatible).

granules(short_name="GEDI02_A")
# output
10-element Vector{EarthData.UMM_G}:
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G
 EarthData.UMM_G

EarthData.netrc! — Method

netrc!(username, password)

Writes/updates a .netrc file for ICESat-2 and GEDI downloads. A .netrc is a plaintext file containing your username and password for NASA EarthData and DAACs, and can be automatically used by Julia using Downloads and tools like wget, curl among others.